Remove dependance on ts-lamdba-handler
