--- deployment/airview-cms-api-aws-lambda-wrapper/src/Api/Cms/Branches/index.ts
@@ -1,2 +1,0 @@
-export { GetBranches } from "./GetBranches.js";
-export { PostBranches } from "./PostBranches.js";
