--- deployment/airview-cms-api-aws-lambda-wrapper/src/Api/Cms/Entries/GetEntryBySha.ts
@@ -1,27 +1,0 @@
-import { CmsApiHandler } from "../../../Handler/CmsApiHandler.js";
-import { Errors, Request, Response, Utilities } from "ts-lambda-handler";
-
-export class GetEntryBySha extends CmsApiHandler {
-  public async process(request: Request, response: Response): Promise<void> {
-    let data: any;
-    try {
-      data = await this.cmsBackend.getEntries(request.getPathParameter("sha"));
-      Utilities.Functions.print_debug(`Data: ${data}`);
-    } catch (error) {
-      console.log(error);
-      this.response.fail(
-        new Errors.InternalServerError({
-          Function: this.context.functionName,
-          Name: this.context.logStreamName,
-          Request: this.context.awsRequestId,
-        })
-      );
-    }
-
-    response
-      .setBody(JSON.stringify(data))
-      .addHeader("content-type", "application/json")
-      .send();
-    return Promise.resolve();
-  }
-}
