--- deployment/airview-cms-api-aws-lambda-wrapper/src/Api/Cms/Entries/index.ts
@@ -1,1 +1,0 @@
-export { GetEntryBySha } from "./GetEntryBySha.js";
