--- deployment/airview-cms-api-aws-lambda-wrapper/src/Api/Cms/Media/index.ts
@@ -1,1 +1,0 @@
-export { GetMediaBySha } from "./GetMediaBySha.js";
