--- deployment/airview-cms-api-aws-lambda-wrapper/src/Api/Cms/Pulls/index.ts
@@ -1,1 +1,0 @@
-export { PostPulls } from "./PostPulls";
