--- deployment/airview-cms-api-aws-lambda-wrapper/src/Api/Cms/Search/index.ts
@@ -1,1 +1,0 @@
-export { GetSearchBySha } from "./GetSearchBySha";
