--- deployment/airview-cms-api-aws-lambda-wrapper/src/Api/Cms/index.ts
@@ -1,8 +1,0 @@
-import * as Branches from "./Branches";
-import * as Entries from "./Entries";
-import * as Content from "./Content";
-import * as Search from "./Search";
-import * as Pulls from "./Pulls";
-import * as Media from "./Media";
-
-export { Content, Branches, Entries, Search, Pulls, Media };
