--- deployment/airview-cms-api-aws-lambda-wrapper/src/Api/index.ts
@@ -1,3 +1,0 @@
-import * as CMS from "./Cms";
-
-export { CMS };
