--- deployment/airview-cms-api-aws-lambda-wrapper/src/Handler/index.ts
@@ -1,1 +1,0 @@
-export { CmsApiHandler } from "./CmsApiHandler.js";
