--- deployment/airview-cms-api-aws-lambda-wrapper/src/Utilities/index.ts
@@ -1,2 +1,0 @@
-import * as Functions from "./Functions.js";
-export { Functions };
