+++ deployment/airview-cms-api-aws-lambda-wrapper/src/api/putContentEntity.ts
@@ -1,7 +1,53 @@
-import { CmsApiHandler } from "../../../Handler/CmsApiHandler.js";
-import { Errors, Request, Response, Utilities } from "ts-lambda-handler";
+import { Context, APIGatewayProxyResult, APIGatewayEvent } from "aws-lambda";
+import { cmsBackendFactory, utils, getAuthorDetails } from "../common";
 
-export class PutContentEntity extends CmsApiHandler {
+export function putContentEntity() {
+  const factory = cmsBackendFactory();
+  async function handle(
+    event: APIGatewayEvent,
+    context: Context
+  ): Promise<APIGatewayProxyResult> {
+    const cmsBackend = await factory.getInstance();
+    const branch: string = event.queryStringParameters["branch"] || "";
+    const baseSha: string = event.queryStringParameters["baseSha"] || "";
+
+    if (branch === "" || baseSha === "") {
+      console.error("Missing query string parameters!");
+
+      return {
+        statusCode: 400,
+        body: "Bad Request. Missing query params",
+      };
+    }
+
+    const entityId = `${event.pathParameters["collection"]}/${event.pathParameters["entity"]}`;
+    utils.printDebug(`entityID == "${entityId}"`);
+
+    const body = JSON.parse(event.body);
+    utils.printDebug(`body == "${body}"`);
+
+    const cookie = this.event.headers["cookie"];
+    const author = await getAuthorDetails(cookie);
+    await cmsBackend.setContent({
+      id: entityId,
+      branchName: branch,
+      baseSha: baseSha,
+      content: body,
+      author,
+    });
+
+    return {
+      statusCode: 201,
+      body: "Accepted",
+    };
+  }
+
+  return { handle };
+}
+
+/*
+
+export class putContentEntity extends CmsApiHandler {
   public async process(request: Request, response: Response): Promise<void> {
     const branch: string = request.getQueryStringParameter("branch", undefined);
     Utilities.Functions.print_debug(`Query String: branch == "${branch}"`);
@@ -28,7 +74,7 @@ export class PutContentEntity extends CmsApiHandler {
 
     try {
       const cookie = request.getHeader("cookie");
-      const author = await this.getAuthorDetails(cookie);
+t     const author = await this.getAuthorDetails(cookie);
       await this.cmsBackend.setContent({
         id: entityId,
         branchName: branch,
@@ -50,4 +96,6 @@ export class PutContentEntity extends CmsApiHandler {
 
     return Promise.resolve();
   }
+ 
 }
+*/
