+++ deployment/airview-cms-api-aws-lambda-wrapper/src/index.ts
@@ -1,5 +1,1 @@
-export { Request, Response, Utilities, Errors } from "ts-lambda-handler";
-export { ProxyHandler } from "aws-lambda";
-
-import * as API from "./Api";
-export { API };
+export * from "./api";
